+++ ccloud/stop-docker.sh
@@ -15,12 +15,15 @@ for v in $(docker volume ls -q --filter="dangling=true"); do
         docker volume rm "$v"
 done
 
-# Delete subjects from Confluent Schema Registry
-schema_registry_subjects_to_delete="users-value pageviews-value"
-for subject in $schema_registry_subjects_to_delete
-do
-  curl -X DELETE --silent -u $SCHEMA_REGISTRY_BASIC_AUTH_USER_INFO $SCHEMA_REGISTRY_URL/subjects/$subject
-done
+# Delete subjects from Confluent Cloud Schema Registry
+. ./config.sh
+if [[ "${USE_CONFLUENT_CLOUD_SCHEMA_REGISTRY}" == true ]]; then
+  schema_registry_subjects_to_delete="users-value pageviews-value"
+  for subject in $schema_registry_subjects_to_delete
+  do
+    curl -X DELETE --silent -u $SCHEMA_REGISTRY_BASIC_AUTH_USER_INFO $SCHEMA_REGISTRY_URL/subjects/$subject
+  done
+fi
 
 topics=$(ccloud topic list)
